requirement.txt
pyupbit
streamlit
openai
pandas
numpy
sqlalchemy
python-dotenv
